Tests: Sig Mint Arithmetic Overflow/Underflow issue fix

+++ packages/manifold/test/lazyclaim/ERC1155LazyPayableClaimSignatureMinting.t.sol
@@ -87,7 +87,7 @@ contract ERC1155LazyPayableClaimSignatureMintingTest is Test {
       vm.startPrank(other);
 
       bytes32 nonce = "1";
-      uint expiration = block.timestamp + 120;
+      uint expiration = block.timestamp + uint(120);
       bytes32 message = keccak256(abi.encodePacked("\x19Ethereum Signed Message:\n32", uint256(1), nonce, other2, expiration));
 
       (uint8 v, bytes32 r, bytes32 s) = vm.sign(privateKey, message);
@@ -121,7 +121,7 @@ contract ERC1155LazyPayableClaimSignatureMintingTest is Test {
 
       // Bad message signed
       nonce = "2";
-      expiration = block.timestamp + 120;
+      expiration = block.timestamp + uint(120);
       message = keccak256(abi.encodePacked("\x19Ethereum Signed Message:\n32", nonce, uint256(1), other2, expiration));
 
       (v, r, s) = vm.sign(privateKey, message);
@@ -141,7 +141,7 @@ contract ERC1155LazyPayableClaimSignatureMintingTest is Test {
 
       // Correct message, wrong signer
       nonce = "2";
-      expiration = block.timestamp + 120;
+      expiration = block.timestamp + uint(120);
       message = keccak256(abi.encodePacked("\x19Ethereum Signed Message:\n32", uint256(1), nonce, other2, expiration));
       (v, r, s) = vm.sign(privateKey2, message);
       signature = abi.encodePacked(r, s, v);
@@ -159,7 +159,7 @@ contract ERC1155LazyPayableClaimSignatureMintingTest is Test {
 
       // Expired signature
       nonce = "2";
-      expiration = block.timestamp - 60;
+      expiration = block.timestamp - uint(60);
       message = keccak256(abi.encodePacked("\x19Ethereum Signed Message:\n32", uint256(1), nonce, other2, expiration));
       (v, r, s) = vm.sign(privateKey2, message);
       signature = abi.encodePacked(r, s, v);
@@ -203,7 +203,7 @@ contract ERC1155LazyPayableClaimSignatureMintingTest is Test {
       vm.stopPrank();
       vm.startPrank(other);
       nonce = "2";
-      expiration = block.timestamp + 120;
+      expiration = block.timestamp + uint(120);
       message = keccak256(abi.encodePacked("\x19Ethereum Signed Message:\n32", uint256(1), nonce, other2, expiration));
       (v, r, s) = vm.sign(privateKey, message);
       signature = abi.encodePacked(r, s, v);
